pyproject.toml
[tool.poetry]
name = "pollsplus"
version = "0.1.0"
description = ""
authors = ["ece-mohammad <kuro.ece@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"
django-werkzeug = "^1.0.0"
ipython = "^8.27.0"
django-ipython = "^0.1.1"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
pytest-xdist = "^3.6.1"
pytest-sugar = "^1.0.0"
django-test-plus = "^2.2.4"
django-robots = "^6.1"
django-redirects = "^0.6.0"
django-extra-views = "^0.15.0"
django-vanilla-views = "^3.0.0"
neapolitan = "^24.6"
django-filter = "^24.3"
django-crispy-forms = "^2.3"
crispy-bootstrap4 = "^2024.10"
pybase62 = "^1.0.0"
mysqlclient = "^2.2.5"
pillow = "^11.1.0"
pymemcache = "^4.0.0"
django-redis = "^5.4.0"
redis = "^5.2.1"
whitenoise = "^6.8.2"
django-simple-history = "^3.7.0"


[tool.poetry.group.dev.dependencies]
django-test-plus = "^2.2.4"
pytest-django = "^4.9.0"
pytest = "^8.3.3"
pytest-xdist = "^3.6.1"
pytest-sugar = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "pollsplus.settings"
python_files = ["test_*.py", "*_test.py", "tests.py"]
required_plugins = ["pytest-django"]
addopts = "-x -l --no-header --fail-on-template-vars"
filterwarnings = ["ignore::DeprecationWarning"]
